+++ crates/internal/composite-service/src/transformers.rs
@@ -94,6 +94,7 @@ impl ForeignFrom<&CompositeAuthorizeRequest> for CustomerServiceCreateRequest {
             metadata: item.metadata.clone(),
             connector_feature_data: item.connector_feature_data.clone(),
             test_mode: item.test_mode,
+            split_payments: item.split_payments.clone(),
         }
     }
 }
@@ -202,6 +203,8 @@ impl
             l2_l3_data: item.l2_l3_data.clone(),
             connector_order_id: item.connector_order_id.clone(),
             merchant_request_id: item.merchant_request_id.clone(),
+            domain_data: item.domain_data.clone(),
+            split_payments: item.split_payments.clone(),
         }
     }
 }
@@ -267,6 +270,7 @@ impl
             payment_experience: item.payment_experience,
             merchant_request_id: item.merchant_request_id.clone(),
             payment_method_type: item.payment_method_type,
+            split_payments: item.split_payments.clone(),
         }
     }
 }
@@ -335,6 +339,7 @@ impl
             merchant_request_id: item.merchant_request_id.clone(),
             connector_order_id: item.connector_order_id.clone(),
             payment_method: item.payment_method.clone(),
+            split_refunds: item.split_refunds.clone(),
         }
     }
 }
@@ -398,6 +403,7 @@ impl
             refund_amount: item.refund_amount,
             merchant_request_id: item.merchant_request_id.clone(),
             connector_order_id: item.connector_order_id.clone(),
+            split_refunds: item.split_refunds.clone(),
         }
     }
 }
@@ -934,6 +940,8 @@ impl
             l2_l3_data: request.l2_l3_data.clone(),
             connector_order_id: request.connector_order_id.clone(),
             merchant_request_id: request.merchant_request_id.clone(),
+            domain_data: None,
+            split_payments: request.split_payments.clone(),
         }
     }
 }